Se avanzó en editar sprint, se arreglo la base de datos

+++ src/components/ui/SprintCard/SprintCard.tsx
@@ -1,49 +1,65 @@
 import { FC } from "react";
 import { ISprint } from "../../../types/ISprint";
-import { FaEye, FaEdit, FaTrash } from "react-icons/fa"; // Importa los iconos de react-icons
-import Swal from "sweetalert2"; // Importar SweetAlert2
+import { FaEye, FaEdit, FaTrash } from "react-icons/fa";
+import Swal from "sweetalert2";
 import styles from "./SprintCard.module.css";
 
-// Tipos de props
 type IPropsSprintCard = {
   sprint: ISprint;
   onCardClick?: () => void;
-  onDelete?: (id: string) => void; // Nueva prop para manejar la eliminación
-  onView?: (sprint: ISprint) => void; // Nueva prop para manejar la visualización
+  onDelete?: (id: string) => void;
+  onView?: (sprint: ISprint) => void;
   onEdit?: (sprint: ISprint) => void;
 };
 
-const SprintCard: FC<IPropsSprintCard> = ({ sprint, onCardClick, onDelete, onView }) => {
-
+const SprintCard: FC<IPropsSprintCard> = ({
+  sprint,
+  onCardClick,
+  onDelete,
+  onView,
+  onEdit,
+}) => {
   const handleDeleteClick = async () => {
     const result = await Swal.fire({
-      title: '¿Estás seguro?',
+      title: "¿Estás seguro?",
       text: `¿Quieres eliminar el sprint ${sprint.nombre}?`,
-      icon: 'warning',
+      icon: "warning",
       showCancelButton: true,
-      confirmButtonText: 'Sí, eliminar',
-      cancelButtonText: 'Cancelar',
+      confirmButtonText: "Sí, eliminar",
+      cancelButtonText: "Cancelar",
     });
 
-    if (result.isConfirmed) {
-      if (onDelete) {
-        onDelete(sprint.id);
-      }
+    if (result.isConfirmed && onDelete) {
+      onDelete(sprint.id);
     }
   };
 
   return (
     <div className={styles.sprintCard}>
-      <div className={styles.sprintCardHeader} onClick={onCardClick}>{sprint.nombre}</div>
+      <div className={styles.sprintCardHeader} onClick={onCardClick}>
+        {sprint.nombre}
+      </div>
       <div className={styles.sprintCardInfo}>
-        <div>Inicio: {new Date(sprint.inicio).toISOString().split("T")[0]}</div>
-        <div>Cierre: {new Date(sprint.fin).toISOString().split("T")[0]}</div>
+        <div>Inicio: {
+          new Date(sprint.inicio).toLocaleDateString('es-AR', {
+            day: '2-digit',
+            month: '2-digit',
+            year: 'numeric'
+          })
+        }</div>
+        <div>Cierre: {
+          new Date(sprint.fin).toLocaleDateString('es-AR', {
+            day: '2-digit',
+            month: '2-digit',
+            year: 'numeric'
+          })
+        }</div>
       </div>
       <div className={styles.sprintCardButtons}>
         <button onClick={() => onView && onView(sprint)}>
           <FaEye size={29} color="C8B810" />
         </button>
-        <button>
+        <button onClick={() => onEdit && onEdit(sprint)}>
           <FaEdit size={25} color="0D7DD8" />
         </button>
         <button onClick={handleDeleteClick}>
